api/requirements.txt
stripe
psycopg2-binary
